Comments updated, improved efficiency
A truncated comment has been corrected (line 106). Additionally, the runtime when calculating the dataset of each "frame" has been massively improved. Instead of adding up the sum of the coefficients again and again starting from k=0 in "get_fourier_vector_line" (line 53) for every single point (of the moving "Fourier arm"), now the new coefficient is just added to the previous one. This improves the calculation of the "frames" for the animation, which took the most time so far.